requirements.txt
alembic==1.13.0
annotated-types==0.6.0
asyncpg==0.29.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
cachelib==0.10.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
greenlet==3.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.1
psycopg==3.1.14
psycopg2==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
scipy==1.11.3
simple-websocket==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==2.3.0
wsproto==1.2.0
WTForms==3.1.0
